adding exception

+++ src/main/java/com/dhbinh/personalproject/service/impl/DishCategoryServiceImpl.java
@@ -61,6 +61,8 @@ public class DishCategoryServiceImpl implements DishCategoryService {
 
     @Override
     public void deleteByType(String type) {
+        if(!dishCategoryRepository.existsByType(type))
+            throw PersonalProjectException.dishCategoryNotFound();
         dishCategoryRepository.deleteByType(type);
     }
 }
